+++ app/person-escort-record/controllers/index.js
@@ -1,11 +1,5 @@
-const frameworkOverviewController = require('./framework-overview')
-const FrameworkSectionController = require('./framework-section')
-const FrameworkStepController = require('./framework-step')
 const printRecordController = require('./print-record')
 
 module.exports = {
-  frameworkOverviewController,
-  FrameworkSectionController,
-  FrameworkStepController,
   printRecordController,
 }
