+++ app/person-escort-record/middleware.js
@@ -1,35 +1,4 @@
-const frameworksService = require('../../common/services/frameworks')
 const personEscortRecordService = require('../../common/services/person-escort-record')
-
-function setFramework(req, res, next) {
-  if (!req.personEscortRecord) {
-    return next()
-  }
-
-  try {
-    req.framework = frameworksService.getPersonEscortRecord(
-      req.personEscortRecord?.version
-    )
-
-    next()
-  } catch (error) {
-    next(error)
-  }
-}
-
-function setFrameworkSection(req, res, next, key) {
-  const section = req.framework?.sections[key]
-
-  if (section) {
-    req.frameworkSection = section
-    return next()
-  }
-
-  const error = new Error('Framework section not found')
-  error.statusCode = 404
-
-  next(error)
-}
 
 async function setPersonEscortRecord(req, res, next) {
   const recordId = req.params?.personEscortRecordId
@@ -54,7 +23,5 @@ async function setPersonEscortRecord(req, res, next) {
 }
 
 module.exports = {
-  setFramework,
-  setFrameworkSection,
   setPersonEscortRecord,
 }
