+++ common/controllers/framework/index.js
@@ -1,0 +1,9 @@
+const frameworkOverviewController = require('./framework-overview')
+const FrameworkSectionController = require('./framework-section')
+const FrameworkStepController = require('./framework-step')
+
+module.exports = {
+  frameworkOverviewController,
+  FrameworkSectionController,
+  FrameworkStepController,
+}
