+++ common/middleware/framework/index.js
@@ -1,0 +1,7 @@
+const setFramework = require('./set-framework')
+const setFrameworkSection = require('./set-framework-section')
+
+module.exports = {
+  setFramework,
+  setFrameworkSection,
+}
