+++ common/templates/framework-section.njk
